Kimi redesign/myaitutor/mait-mvp-glitch/backend/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
pydantic
python-multipart
requests
google-genai
python-dotenv
aiosqlite
# RAG dependencies
sentence-transformers
pypdf
python-docx
# Testing dependencies
pytest
httpx
pytest-asyncio
slowapi
google-auth
sentry-sdk
